Move rotate module.
