+++ dolfyn/adv/api.py
@@ -37,6 +37,6 @@ Examples
 
 from .turbulence import calc_turbulence, TurbBinner
 from . import clean
-from . import rotate
+from ..nortek.rotate import vector as rotate
 from . import motion
 from ..io.api import read, load
