+++ dolfyn/test/data/vector_data01_rotate_inst2beam.h5
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:c1d6f36e456354237fce81bb8d72d60b539efd91869aa43bb0bba90b8b748a7b
-size 159027
+oid sha256:a9b75459dbbb1148f153e2a83cf1ae62f833b3588cdfcc3dab4db4531a4b767d
+size 158947
