+++ dolfyn/test/data/vector_data_imu01_rotate_inst2beam.h5
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:55563c461a94b6c1fb4907fb8828e5c8735d765fc016be8d3c124c3a412f3bae
+oid sha256:82ca8e7e77ccc14f3eaab428432ca677f1b6c7e162b367de0c6080857008e0df
 size 185469
